+++ src/websocket/node.ts
@@ -1,6 +1,5 @@
 import { WebSocket as _WebSocket } from "ws";
 
-const Websocket: typeof globalThis.WebSocket =
-  globalThis.WebSocket || _WebSocket;
+const Websocket: typeof globalThis.WebSocket = globalThis.WebSocket || _WebSocket;
 
 export default Websocket;